requirements.txt
matplotlib==3.3.4
numpy==1.18.1
opencv-python==4.5.3.56
pandas==1.1.5
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
scikit-image==0.17.2
scipy==1.1.0
tensorboard-logger==0.1.0
torch==1.7.0
torchvision==0.8.0